requirements.txt
flask
requests
lazy
python-jose